Fix error in doc and improve I/O exception test case
