+++ deployment/src/test/java/io/quarkiverse/logging/splunk/LoggingSplunkMandatoryConfigTest.java
