+++ deployment/src/test/resources/mock-server.properties
@@ -4,7 +4,9 @@ quarkus.log.handler.splunk.url=http://localhost:8088
 quarkus.log.handler.console.\"stdout\".format=%s%e%n
 quarkus.log.category.\"org.mockserver\".handlers=stdout
 quarkus.log.category.\"org.mockserver\".use-parent-handlers=false
-# Avoid batching and send events immediately, to make unit tests synchronous
+# Avoid batching and send events immediately, to make unit tests more synchronous
+# Note that OKHttp client still executes its I/O on a separate thread
 quarkus.log.handler.splunk.batch-interval=0
 quarkus.log.handler.splunk.batch-size-bytes=0
 quarkus.log.handler.splunk.batch-size-count=0
+quarkus.log.handler.splunk.send-mode=sequential
